+++ CHANGELOG.md
@@ -10,6 +10,8 @@ This is a small bug fix release.
 
 ### Changelog
 
+- Fix: add playlist to queue from playlist view #1087
+
 ***
 
 ## myMPD v11.0.3 (2023-08-15)